エラー仕様の追加

+++ src/app/comment/usecase/delete-comment-usecase.ts
@@ -17,6 +17,9 @@ export class DeletCommentUseCase {
             commentId
         } = params
         await new DomainService().tokenCheck(token);
+        if ('tokenError') {
+            return 'tokenError'
+        }
 
         await this.commentRepo.delete(postId, userId, commentId);
